CRDCDH-575 moving icons (#203)
* moving icons

* update test case imports

* renaming the updated icons
