+++ src/assets/history/submissionRequest/index.tsx
@@ -1,5 +1,5 @@
-import New from "./New.svg";
-import Submitted from "./Submitted.svg";
+import New from "./SubmissionRequestNew.svg";
+import Submitted from "./SubmissionRequestSubmitted.svg";
 import Rejected from "./Rejected.svg";
 import Approved from "./Approved.svg";
 import UnderReview from "./UnderReview.svg";
